Use proper APIs and remove unnecessary conditions
The old round-robin scheduler was causing some test to fail
when not using the mlfqs. Also moved some checks around.

+++ pintos/pintos-env/pintos/threads/thread.h
@@ -100,7 +100,7 @@ struct thread
 #endif
 
     int64_t wakeup_tick;                /* Tick at which to wake the thread up. */
-    int8_t nice;                        /* "Nice" value for scheduler */
+    int nice;                           /* "Nice" value for scheduler */
     FPReal recent_cpu;                  /* How much cpu the thread has used recently */
 
     /* Owned by thread.c. */
